+++ src/main/java/com/example/linter/config/blocks/AbstractBlock.java
@@ -2,7 +2,7 @@ package com.example.linter.config.blocks;
 
 import com.example.linter.config.BlockType;
 import com.example.linter.config.Severity;
-import com.example.linter.rule.OccurrenceRule;
+import com.example.linter.config.rule.OccurrenceRule;
 
 import java.util.Objects;
