+++ src/main/java/com/example/linter/config/blocks/ListingBlock.java
@@ -2,7 +2,7 @@ package com.example.linter.config.blocks;
 
 import com.example.linter.config.BlockType;
 import com.example.linter.config.Severity;
-import com.example.linter.rule.LineRule;
+import com.example.linter.config.rule.LineRule;
 
 import java.util.*;
 import java.util.regex.Pattern;
