+++ src/main/java/com/example/linter/config/rule/AttributeRule.java
@@ -1,4 +1,4 @@
-package com.example.linter.rule;
+package com.example.linter.config.rule;
 
 import com.example.linter.config.Severity;
